# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Campus Portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OkResponse {
  ok: boolean;
}

export type Role = (typeof Role)[keyof typeof Role];

export const Role = {
  student: "student",
  faculty: "faculty",
  admin: "admin",
} as const;

export interface LoginBody {
  role: Role;
  /** Roll number for student, faculty ID for faculty, email for admin */
  identifier: string;
  password: string;
}

export interface AuthSession {
  authenticated: boolean;
  userId?: string | null;
  role?: string | null;
  name?: string | null;
  identifier?: string | null;
  mustChangePassword: boolean;
}

export interface ChangePasswordBody {
  currentPassword: string;
  newPassword: string;
}

export interface Student {
  id: string;
  rollNumber: string;
  name: string;
  email?: string | null;
  department?: string | null;
  groupId?: string | null;
  passwordChanged: boolean;
}

export interface Faculty {
  id: string;
  facultyId: string;
  name: string;
  email?: string | null;
  department?: string | null;
  groupCount: number;
  passwordChanged: boolean;
}

export interface Domain {
  id: string;
  name: string;
  description?: string | null;
  facultyId: string;
  facultyName: string;
}

export interface FacultyWithDomains {
  id: string;
  facultyId: string;
  name: string;
  department?: string | null;
  groupCount: number;
  domains: Domain[];
}

export interface Admin {
  id: string;
  name: string;
  email: string;
  createdAt: string;
}

export interface CreateAdminBody {
  name: string;
  email: string;
  password: string;
}

export interface CreateDomainBody {
  name: string;
  description?: string | null;
}

export type GroupStatus = (typeof GroupStatus)[keyof typeof GroupStatus];

export const GroupStatus = {
  pending: "pending",
  assigned: "assigned",
} as const;

export interface Group {
  id: string;
  projectTitle: string;
  description?: string | null;
  facultyId: string;
  facultyName: string;
  domainId: string;
  domainName: string;
  status: GroupStatus;
  createdAt: string;
  members: Student[];
}

export interface CreateGroupBody {
  projectTitle: string;
  description?: string | null;
  memberRollNumbers: string[];
  facultyId: string;
  domainId: string;
}

export interface SheetsConfig {
  studentSheetUrl?: string | null;
  facultySheetUrl?: string | null;
  lastSyncedAt?: string | null;
  lastSyncStatus?: string | null;
  autoSyncEnabled: boolean;
}

export interface UpdateSheetsConfigBody {
  studentSheetUrl?: string | null;
  facultySheetUrl?: string | null;
  autoSyncEnabled?: boolean | null;
}

export interface SyncResult {
  ok: boolean;
  studentsAdded: number;
  studentsRemoved: number;
  facultiesAdded: number;
  facultiesRemoved: number;
  message: string;
}

export interface SecuritySettings {
  sessionTimeoutMinutes: number;
  minPasswordLength: number;
  requirePasswordChange: boolean;
  allowStudentLogin: boolean;
  allowFacultyLogin: boolean;
}

export interface UpdateSecuritySettingsBody {
  sessionTimeoutMinutes?: number | null;
  minPasswordLength?: number | null;
  requirePasswordChange?: boolean | null;
  allowStudentLogin?: boolean | null;
  allowFacultyLogin?: boolean | null;
}

export interface AuditLog {
  id: string;
  actorRole: string;
  actorName: string;
  action: string;
  details?: string | null;
  createdAt: string;
}

export type MarkAttendanceBodyEntriesItemStatus =
  (typeof MarkAttendanceBodyEntriesItemStatus)[keyof typeof MarkAttendanceBodyEntriesItemStatus];

export const MarkAttendanceBodyEntriesItemStatus = {
  present: "present",
  absent: "absent",
  late: "late",
} as const;

export type MarkAttendanceBodyEntriesItem = {
  studentId: string;
  status: MarkAttendanceBodyEntriesItemStatus;
};

export interface MarkAttendanceBody {
  date: string;
  entries: MarkAttendanceBodyEntriesItem[];
}

export interface AttendanceEntry {
  id: string;
  studentId: string;
  studentName: string;
  rollNumber: string;
  date: string;
  status: string;
}

export interface AttendanceSummary {
  totalSessions: number;
  presentCount: number;
  absentCount: number;
  lateCount: number;
  attendancePercent: number;
  entries: AttendanceEntry[];
}

export interface SendEmailBody {
  recipientStudentIds: string[];
  subject: string;
  body: string;
}

export interface EmailLog {
  id: string;
  subject: string;
  body: string;
  recipientCount: number;
  recipientNames: string[];
  sentAt: string;
}

export interface Notification {
  id: string;
  title: string;
  body: string;
  read: boolean;
  createdAt: string;
}

export type RecordPerformanceBodyPeriod =
  (typeof RecordPerformanceBodyPeriod)[keyof typeof RecordPerformanceBodyPeriod];

export const RecordPerformanceBodyPeriod = {
  weekly: "weekly",
  monthly: "monthly",
  semester: "semester",
} as const;

export interface RecordPerformanceBody {
  studentId: string;
  score: number;
  notes?: string | null;
  period: RecordPerformanceBodyPeriod;
}

export interface PerformanceEntry {
  id: string;
  studentId: string;
  studentName: string;
  rollNumber: string;
  score: number;
  notes?: string | null;
  period: string;
  recordedAt: string;
}

export type PerformanceReportTopPerformer = {
  studentId: string;
  name: string;
  rollNumber: string;
  score: number;
} | null;

export type PerformanceReportAttendanceByStudentItem = {
  studentId: string;
  name: string;
  rollNumber: string;
  attendancePercent: number;
};

export interface PerformanceReport {
  period: string;
  averageScore: number;
  topPerformer?: PerformanceReportTopPerformer;
  entries: PerformanceEntry[];
  attendanceByStudent: PerformanceReportAttendanceByStudentItem[];
}

export type AdminDashboardGroupsByDomainItem = {
  domain: string;
  count: number;
};

export interface AdminDashboard {
  studentCount: number;
  facultyCount: number;
  adminCount: number;
  groupCount: number;
  domainsCount: number;
  attendanceRate: number;
  lastSyncedAt?: string | null;
  lastSyncStatus?: string | null;
  recentActivity: AuditLog[];
  groupsByDomain: AdminDashboardGroupsByDomainItem[];
}

export interface FacultyDashboard {
  groupCount: number;
  groupCapacity: number;
  studentCount: number;
  domainsCount: number;
  attendanceRate: number;
  unreadNotifications: number;
  recentEmails: EmailLog[];
  upcomingGroups: Group[];
}

export interface StudentDashboard {
  hasGroup: boolean;
  group?: Group | null;
  attendancePercent: number;
  averagePerformance: number;
  unreadNotifications: number;
  recentNotifications: Notification[];
}

export type GetAttendanceForDateParams = {
  date: string;
};

export type GetPerformanceReportParams = {
  period: GetPerformanceReportPeriod;
  groupId?: string;
};

export type GetPerformanceReportPeriod =
  (typeof GetPerformanceReportPeriod)[keyof typeof GetPerformanceReportPeriod];

export const GetPerformanceReportPeriod = {
  weekly: "weekly",
  monthly: "monthly",
  semester: "semester",
} as const;