# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Campus Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginBody" }
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthSession" }
...
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChangePasswordBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }

  # ===== ADMIN =====
...
  /admin/sheets:
    get:
      operationId: getSheetsConfig
      tags: [admin]
      responses:
        "200":
          description: Sheets config
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SheetsConfig" }
    put:
      operationId: updateSheetsConfig
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateSheetsConfigBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SheetsConfig" }
...
  /admin/admins:
    get:
      operationId: listAdmins
      tags: [admin]
      responses:
        "200":
          description: Admins
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Admin" }
    post:
      operationId: createAdmin
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateAdminBody" }
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Admin" }
  /admin/admins/{adminId}:
    delete:
      operationId: deleteAdmin
      tags: [admin]
      parameters:
        - name: adminId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
  /admin/security:
    get:
      operationId: getSecuritySettings
      tags: [admin]
      responses:
        "200":
          description: Security settings
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SecuritySettings" }
    put:
      operationId: updateSecuritySettings
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateSecuritySettingsBody" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SecuritySettings" }
...
components:
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
    Role:
      type: string
      enum: [student, faculty, admin]
    LoginBody:
      type: object
      properties:
        role: { $ref: "#/components/schemas/Role" }
        identifier: { type: string, description: "Roll number for student, faculty ID for faculty, email for admin" }
        password: { type: string }
      required: [role, identifier, password]
    AuthSession:
      type: object
      properties:
        authenticated: { type: boolean }
        userId: { type: string, nullable: true }
        role: { type: string, nullable: true }
        name: { type: string, nullable: true }
        identifier: { type: string, nullable: true }
        mustChangePassword: { type: boolean }
      required: [authenticated, mustChangePassword]
    ChangePasswordBody:
      type: object
      properties:
        currentPassword: { type: string }
        newPassword: { type: string }
      required: [currentPassword, newPassword]
...
    Admin:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        createdAt: { type: string }
      required: [id, name, email, createdAt]
    CreateAdminBody:
      type: object
      properties:
        name: { type: string }
        email: { type: string }
        password: { type: string }
      required: [name, email, password]
...
    SheetsConfig:
      type: object
      properties:
        studentSheetUrl: { type: string, nullable: true }
        facultySheetUrl: { type: string, nullable: true }
        lastSyncedAt: { type: string, nullable: true }
        lastSyncStatus: { type: string, nullable: true }
        autoSyncEnabled: { type: boolean }
      required: [autoSyncEnabled]
    UpdateSheetsConfigBody:
      type: object
      properties:
        studentSheetUrl: { type: string, nullable: true }
        facultySheetUrl: { type: string, nullable: true }
        autoSyncEnabled: { type: boolean, nullable: true }
...
    SecuritySettings:
      type: object
      properties:
        sessionTimeoutMinutes: { type: integer }
        minPasswordLength: { type: integer }
        requirePasswordChange: { type: boolean }
        allowStudentLogin: { type: boolean }
        allowFacultyLogin: { type: boolean }
      required: [sessionTimeoutMinutes, minPasswordLength, requirePasswordChange, allowStudentLogin, allowFacultyLogin]
    UpdateSecuritySettingsBody:
      type: object
      properties:
        sessionTimeoutMinutes: { type: integer, nullable: true }
        minPasswordLength: { type: integer, nullable: true }
        requirePasswordChange: { type: boolean, nullable: true }
        allowStudentLogin: { type: boolean, nullable: true }
        allowFacultyLogin: { type: boolean, nullable: true }
...