# lib/api-zod/src/generated/types/admin.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Campus Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface Admin {
  id: string;
  name: string;
  email: string;
  createdAt: string;
}