# lib/api-zod/src/generated/types/authSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Campus Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface AuthSession {
  authenticated: boolean;
  userId?: string | null;
  role?: string | null;
  name?: string | null;
  identifier?: string | null;
  mustChangePassword: boolean;
}