# lib/api-zod/src/generated/types/sheetsConfig.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Campus Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface SheetsConfig {
  studentSheetUrl?: string | null;
  facultySheetUrl?: string | null;
  lastSyncedAt?: string | null;
  lastSyncStatus?: string | null;
  autoSyncEnabled: boolean;
}