# lib/api-zod/src/generated/types/updateSheetsConfigBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Campus Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateSheetsConfigBody {
  studentSheetUrl?: string | null;
  facultySheetUrl?: string | null;
  autoSyncEnabled?: boolean | null;
}